// src/modules/quran-audio/enums/reciter.enum.ts
export enum QuranReciter {
  ABDUL_SAMAD = 'ar.abdulsamad',
  HUSARY = 'ar.husary',
  HUSARY_MUJAWWAD = 'ar.husarymujawwad',
  MUHAMMAD_JIBREEL = 'ar.muhammadjibreel',
  MINSHAWI = 'ar.minshawi',
  MINSHAWI_MUJAWWAD = 'ar.minshawimujawwad'
}

export enum QuranReciterArabic {
  ABDUL_SAMAD = 'عبد الباسط عبد الصمد',
  HUSARY = 'محمود خليل الحصري',
  HUSARY_MUJAWWAD = 'محمود خليل الحصري - مجود',
  MUHAMMAD_JIBREEL = 'محمد جبريل',
  MINSHAWI = 'محمد صديق المنشاوي',
  MINSHAWI_MUJAWWAD = 'محمد صديق المنشاوي - مجود'
}

export enum QuranReciterEnglish {
  ABDUL_SAMAD = 'Abdul Basit Abdul Samad',
  HUSARY = 'Mahmoud Khalil Al-Hussary',
  HUSARY_MUJAWWAD = 'Mahmoud Khalil Al-Hussary (Mujawwad)',
  MUHAMMAD_JIBREEL = 'Muhammad Jibreel',
  MINSHAWI = 'Mohamed Siddiq Al-Minshawi',
  MINSHAWI_MUJAWWAD = 'Mohamed Siddiq Al-Minshawi (Mujawwad)'
}